+++ src/app/admin/categories/CategoriesPageClient.tsx
@@ -1,211 +1,7 @@
 'use client';
 
-import React, {useState} from 'react';
-
-import {Tab, Tabs} from '@heroui/tabs';
-
-import {useQuery} from '@tanstack/react-query';
-
-import {useModal, useModalWithItem} from '@/hooks/shared/useModals';
-
-import {translations} from '@/lib/translations';
-
-import {getAgeGroupLabel, getGenderLabel, getStatusClasses, getStatusLabel} from '@/helpers/ui';
-
-import CategoryFeesTab from '@/app/admin/categories/components/CategoryFeesTab';
-
-import {AdminContainer, Dialog, Search, UnifiedTable} from '@/components';
-import {ActionTypes, ModalMode} from '@/enums';
-import {useCategories, useCategoryFiltering, useCategoryForm} from '@/hooks';
-import {fetchCategories} from '@/queries/categories/queries';
-import {Category} from '@/types';
-
-import CategoryModal from './components/CategoryModal';
-
-const t = translations.categories;
-const tAction = translations.common.actions;
+import {CategoriesContainer} from "@/features/categories";
 
 export function CategoriesPageClient() {
-  const {
-    data = [],
-    isLoading: loading,
-    refetch,
-  } = useQuery({
-    queryKey: ['categories'],
-    queryFn: fetchCategories,
-  });
-
-  const [searchTerm, setSearchTerm] = useState('');
-  const {filteredData: categories} = useCategoryFiltering(data, {searchTerm});
-  const {loading: crudLoading, createCategory, updateCategory, deleteCategory} = useCategories();
-
-  const {
-    formData,
-    setFormData,
-    selectedItem: selectedCategory,
-    modalMode,
-    openAddMode,
-    openEditMode,
-    validateForm,
-    resetForm,
-  } = useCategoryForm();
-
-  // Modal states
-  const modal = useModal();
-  const deleteModal = useModalWithItem<Category>();
-
-  const handleAddCategory = () => {
-    openAddMode();
-    modal.onOpen();
-  };
-
-  const handleEdit = async (category: Category) => {
-    openEditMode(category);
-    modal.onOpen();
-  };
-
-  // Enhanced handlers that include modal management
-  const handleSubmit = async () => {
-    const {valid, errors} = validateForm();
-
-    if (!valid) {
-      console.error('Validation errors', errors);
-      return;
-    }
-
-    try {
-      if (modalMode === ModalMode.EDIT && selectedCategory) {
-        await updateCategory(selectedCategory.id, formData);
-      } else {
-        await createCategory(formData);
-      }
-      await refetch();
-      modal.onClose();
-      resetForm();
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const handleDelete = async (category: Category) => {
-    deleteModal.openWith(category);
-  };
-
-  const handleConfirmDelete = async () => {
-    if (!deleteModal.selectedItem) return;
-
-    const success = await deleteCategory(deleteModal.selectedItem.id);
-
-    if (success) {
-      await refetch();
-      deleteModal.closeAndClear();
-    }
-  };
-
-  const categoryColumns = [
-    {key: 'name', label: t.table.name},
-    {key: 'description', label: t.table.description},
-    {key: 'age_group', label: t.table.ageGroup},
-    {key: 'gender', label: t.table.gender},
-    {key: 'is_active', label: t.table.status},
-    {key: 'sort_order', label: t.table.sortOrder},
-    {
-      key: 'actions',
-      label: t.table.actions,
-      isActionColumn: true,
-      actions: [
-        {type: ActionTypes.UPDATE, onPress: handleEdit, title: tAction.edit},
-        {type: ActionTypes.DELETE, onPress: handleDelete, title: tAction.delete},
-      ],
-    },
-  ];
-
-  const renderCategoryCell = (category: Category, columnKey: string) => {
-    switch (columnKey) {
-      case 'name':
-        return <span className="font-medium">{category.name}</span>;
-      case 'description':
-        return <span className="font-medium">{category.description || '-'}</span>;
-      case 'age_group':
-        return <span className="font-medium">{getAgeGroupLabel(category.age_group)}</span>;
-      case 'gender':
-        return <span className="font-medium">{getGenderLabel(category.gender)}</span>;
-      case 'is_active':
-        return (
-          <span className={`font-medium ${getStatusClasses(category.is_active ?? false)}`}>
-            {getStatusLabel(category.is_active ?? false)}
-          </span>
-        );
-      case 'sort_order':
-        return <span className="font-medium">{category.sort_order}</span>;
-    }
-  };
-
-  const filters = (
-    <div className="w-full max-w-md">
-      <Search
-        value={searchTerm}
-        onChange={setSearchTerm}
-        placeholder={translations.categories.placeholders.searchCategory}
-        size="sm"
-      />
-    </div>
-  );
-
-  return (
-    <>
-      <AdminContainer
-        isLoading={loading}
-        filters={filters}
-        actions={[
-          {
-            label: translations.categories.actions.add,
-            onClick: handleAddCategory,
-            variant: 'solid',
-            buttonType: ActionTypes.CREATE,
-          },
-        ]}
-      >
-        <Tabs aria-label="Categories admin tabs">
-          <Tab key="categories" title={translations.categories.page.title}>
-            <UnifiedTable
-              columns={categoryColumns}
-              data={categories}
-              ariaLabel={translations.categories.ariaLabels.table}
-              renderCell={renderCategoryCell}
-              getKey={(category: Category) => category.id}
-              emptyContent={t.table.noCategories}
-              isStriped
-            />
-          </Tab>
-          <Tab key="membershipFees" title={t.modal.membershipFeesTab}>
-            <CategoryFeesTab />
-          </Tab>
-        </Tabs>
-      </AdminContainer>
-
-      <CategoryModal
-        isOpen={modal.isOpen}
-        onClose={modal.onClose}
-        onSubmit={handleSubmit}
-        formData={formData}
-        setFormData={setFormData}
-        mode={modalMode}
-        isLoading={crudLoading}
-      />
-
-      <Dialog
-        isOpen={deleteModal.isOpen}
-        onClose={deleteModal.closeAndClear}
-        title={translations.categories.modal.delete.title}
-        onSubmit={handleConfirmDelete}
-        submitButtonLabel={translations.common.actions.delete}
-        isLoading={crudLoading}
-        dangerAction
-        size={'sm'}
-      >
-        {translations.categories.modal.delete.message}
-      </Dialog>
-    </>
-  );
+ return <CategoriesContainer />
 }
