--- src/app/admin/categories/components/index.ts
@@ -1,1 +1,0 @@
-export {default as CategoryModal} from './CategoryModal';
