--- src/app/admin/categories/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './genderBadgeColor';
-export * from './ageGroupBadgeColor';
