+++ src/components/ui/VStack/VStack.tsx
@@ -6,3 +6,5 @@ export interface VStackProps extends Omit<StackProps, 'direction'> {
 export function VStack(props: VStackProps) {
 	return <Stack direction="column" {...props} />;
 }
+
+VStack.displayName = 'VStack';
