+++ src/features/categories/index.ts
@@ -1,0 +1,1 @@
+export {CategoriesContainer} from './containers/CategoriesContainer';
