+++ src/shared/components/layout/AppPageLayout.tsx
@@ -9,114 +9,109 @@ import {LoadingSpinner, VStack} from '@/components';
 
 import {commonCopy} from '../../copy';
 
-export interface TabConfig {
-  key: string;
-  title: ReactNode;
-  content: ReactNode;
+export interface TabConfig<T extends string = string> {
+	key: T;
+	title: ReactNode;
+	content: ReactNode;
 
-  actions?: ReactNode;
-  filters?: ReactNode;
+	actions?: ReactNode;
+	filters?: ReactNode;
+	floatingActions?: ReactNode;
 
-  inheritGlobalActions?: boolean;
-  inheritGlobalFilters?: boolean;
+	inheritGlobalActions?: boolean;
+	inheritGlobalFilters?: boolean;
 }
 
-export interface AppPageLayoutProps {
-  header?: ReactNode;
-  actions?: ReactNode;
-  filters?: ReactNode;
-  floatingActions?: ReactNode;
+export interface AppPageLayoutProps<T extends string = string> {
+	header?: ReactNode;
+	actions?: ReactNode;
+	filters?: ReactNode;
+	floatingActions?: ReactNode;
 
-  isLoading?: boolean;
-  isError?: boolean;
-  isUnderConstruction?: boolean;
+	isLoading?: boolean;
+	isError?: boolean;
+	isUnderConstruction?: boolean;
 
-  tabs?: TabConfig[];
-  activeTab?: string;
-  onTabChange?: (key: string) => void;
-  tabsAriaLabel?: string;
+	tabs?: TabConfig<T>[];
+	activeTab?: T;
+	onTabChange?: (key: T) => void;
+	tabsAriaLabel?: string;
 
-  children: ReactNode;
+	children?: ReactNode;
 }
 
-export function AppPageLayout({
-  header,
-  actions,
-  filters,
-  floatingActions,
-  isLoading,
-  isError,
-  isUnderConstruction,
+export function AppPageLayout<T extends string = string>({
+															 header,
+															 actions,
+															 filters,
+															 floatingActions,
+															 isLoading,
+															 isError,
+															 isUnderConstruction,
 
-  tabs,
-  activeTab,
-  onTabChange,
-  tabsAriaLabel = 'Tabs',
+															 tabs,
+															 activeTab,
+															 onTabChange,
+															 tabsAriaLabel = 'Tabs',
 
-  children,
-}: AppPageLayoutProps) {
-  if (isLoading) return <LoadingSpinner />;
+															 children,
+														 }: AppPageLayoutProps<T>) {
+	if (isLoading) return <LoadingSpinner/>;
 
-  const effectiveActiveTab = activeTab || tabs?.[0]?.key;
+	const effectiveActiveTab = activeTab || tabs?.[0]?.key;
 
-  const renderTabContent = (tab: TabConfig) => {
-    const tabActions = tab.inheritGlobalActions ? actions : (tab.actions ?? undefined);
+	const activeTabConfig = tabs?.find((t) => t.key === effectiveActiveTab);
 
-    const tabFilters = tab.inheritGlobalFilters ? filters : (tab.filters ?? undefined);
+	const resolvedFilters =
+		activeTabConfig?.filters ??
+		filters;
 
-    return (
-      <VStack spacing={4}>
-        {tabActions}
-        {tabFilters}
-        {tab.content}
-      </VStack>
-    );
-  };
+	const resolvedFloatingActions =
+		activeTabConfig?.floatingActions ??
+		floatingActions;
 
-  return (
-    <>
-      <VStack spacing={4}>
-        {isUnderConstruction && (
-          <Alert
-            color="warning"
-            title={commonCopy.alerts.warning}
-            description={commonCopy.alerts.underConstruction}
-          />
-        )}
-        {isError && (
-          <Alert
-            color="danger"
-            title={commonCopy.alerts.error}
-            description={commonCopy.alerts.errorLoadingPage}
-          />
-        )}
+	return (
+		<>
+			<VStack spacing={4} align={'stretch'}>
+				{isUnderConstruction && (
+					<Alert
+						color="warning"
+						title={commonCopy.alerts.warning}
+						description={commonCopy.alerts.underConstruction}
+					/>
+				)}
+				{isError && (
+					<Alert
+						color="danger"
+						title={commonCopy.alerts.error}
+						description={commonCopy.alerts.errorLoadingPage}
+					/>
+				)}
 
-        {header && header}
+				{header}
 
-        {!tabs && (
-          <>
-            {actions && actions}
-            {filters && filters}
-            {children}
-          </>
-        )}
+				{resolvedFilters}
 
-        {tabs && (
-          <Tabs
-            selectedKey={effectiveActiveTab}
-            onSelectionChange={(key) => onTabChange?.(key as string)}
-            aria-label={tabsAriaLabel}
-          >
-            {tabs.map((tab) => (
-              <Tab key={tab.key} title={tab.title}>
-                {renderTabContent(tab)}
-              </Tab>
-            ))}
-          </Tabs>
-        )}
-      </VStack>
+				{!tabs && children}
 
-      {floatingActions}
-    </>
-  );
+				{tabs && (
+					<Tabs
+						selectedKey={effectiveActiveTab}
+						onSelectionChange={(key) => onTabChange?.(key as T)}
+						aria-label={tabsAriaLabel}
+					>
+						{tabs.map((tab) => (
+							<Tab key={tab.key} title={tab.title}>
+								<VStack spacing={4} align="stretch">
+									{tab.content}
+								</VStack>
+							</Tab>
+						))}
+					</Tabs>
+				)}
+			</VStack>
+
+			{resolvedFloatingActions}
+		</>
+	);
 }
